♻ Refactoring clientside code with mixins
