+++ common/src/main/java/rbasamoyai/createbigcannons/cannon_control/contraption/AbstractMountedCannonContraption.java
@@ -2,10 +2,6 @@ package rbasamoyai.createbigcannons.cannon_control.contraption;
 
 import com.simibubi.create.content.contraptions.components.structureMovement.AssemblyException;
 import com.simibubi.create.content.contraptions.components.structureMovement.Contraption;
-import com.simibubi.create.content.contraptions.components.structureMovement.NonStationaryLighter;
-import com.simibubi.create.content.contraptions.components.structureMovement.render.ContraptionLighter;
-import net.fabricmc.api.EnvType;
-import net.fabricmc.api.Environment;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.nbt.CompoundTag;
@@ -93,9 +89,6 @@ public abstract class AbstractMountedCannonContraption extends Contraption  {
 	public boolean canBeFiredOnController(ControlPitchContraption control) { return true; }
 
 	public BlockPos getSeatPos(Entity entity) { return null; }
-
-	@Environment(EnvType.CLIENT)
-	@Override public ContraptionLighter<?> makeLighter() { return new NonStationaryLighter<>(this); }
 
 	public abstract Vec3 getInteractionVec(PitchOrientedContraptionEntity poce);
